add: bleu benchmark

+++ src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
-    for i in 1..=4 {
+    for i in 1..=4 as usize {
         for j in 0..=(2 - i) {
             println!("{i}, {j}");
         }
